flask/requirements.txt
requests
Pillow
ansicolors
numpy
pysocks
stem

flask
flask_wtf
wtforms
ansi2html